Update README.md
The Soma mobile application aims to connect students seeking sponsorship, sponsors, and schools to facilitate direct support for education in Africa. The project idea is to create a platform that goes beyond traditional scholarship models, empowering the community to directly impact the educational journey of students, ensuring that financial barriers do not hinder their potential. A platform where sponsors be it individuals and organization will be able to select a student and pay for their needs be it tuition, accommodation bills or study materials.